+++ @navikt/aksel/src/codemod/migrations.ts
@@ -97,35 +97,31 @@ export const migrations: {
         "Remember to update use of `variant`-prop to match previous use of colors. If needed the component exposes css-variables for custom overrides",
     },
   ],
-  darkside: [
+  spacing: [
     {
       description:
         "Updates all Primitives to use new `space`-tokens. (Works with old and new system)",
       value: "primitive-spacing",
-      path: "darkside/primitives-spacing/spacing",
+      path: "spacing/primitives-spacing/spacing",
     },
     {
       description:
         "Updates css, scss and less-variables to use new `space`-tokens. (Works with old and new system)",
       value: "token-spacing",
-      path: "darkside/token-spacing/spacing",
+      path: "spacing/token-spacing/spacing",
     },
     {
       description:
         "Updates js-tokens to use new `space`-tokens. (Works with old and new system)",
       value: "token-spacing-js",
-      path: "darkside/token-spacing-js/spacing",
+      path: "spacing/token-spacing-js/spacing",
     },
+  ],
+  darkside: [
     {
-      description:
-        "Updates css, scss and less-variables to use new token-prefix",
-      value: "token-update",
-      path: "darkside/token-update/token-update",
-    },
-    {
-      description: "Updates js tokens to use new token-prefix and names",
-      value: "token-update-js",
-      path: "darkside/token-update-js/token-update-js",
+      description: "marks deprecated prop usage with comments.",
+      value: "prop-deprecate",
+      path: "darkside/prop-deprecate/prop-deprecate",
     },
   ],
 };