+++ aksel.nav.no/website/app/dev/(designsystemet)/komponenter/page.tsx
@@ -1,0 +1,41 @@
+/* import { client } from "@/app/_sanity/client"; */
+import { DesignsystemetPageLayout } from "../_ui/DesignsystemetPageLayout";
+
+/* import { disableDraftMode, enableDraftMode } from "@/app/actions"; */
+
+/* https://nextjs.org/docs/app/api-reference/file-conventions/page#props */
+export default async function Page(/* {
+  params,
+}: {
+  params: Promise<{ slug: string }>;
+} */) {
+  /* const data = await client.fetch(
+    "*[_type == 'komponent_artikkel'][0].heading",
+    undefined,
+    { useCdn: false, perspective: "drafts" },
+  ); */
+
+  /* <button onClick={enableDraftMode}>on</button>
+  <hr />
+  <button onClick={disableDraftMode}>off</button>
+
+  <pre>{data}</pre> */
+
+  return (
+    <DesignsystemetPageLayout>
+      <div style={{ background: "red", width: "100%", height: 50 }}></div>
+      <div style={{ background: "blue", width: "240px", height: 50 }}></div>
+    </DesignsystemetPageLayout>
+  );
+}
+
+/*
+
+*[_type == "komponent_artikkel"][2]{
+  "TOC": content[style match 'h2'][]{
+    _key,
+    "text": pt::text(@)
+  }
+}
+
+*/