+++ aksel.nav.no/website/app/theme.d.ts
@@ -1,0 +1,13 @@
+import "react";
+import type { GlobalColorRoles } from "@navikt/ds-tokens/types";
+
+type AkselBrandColors =
+  | "aksel-brand-teal"
+  | "aksel-brand-violet"
+  | "aksel-brand-pink";
+
+declare module "react" {
+  interface HTMLAttributes {
+    "data-color-role"?: GlobalColorRoles | AkselBrandColors;
+  }
+}