+++ aksel.nav.no/website/tailwind.config.js
@@ -5,6 +5,7 @@ module.exports = {
     "./sanity/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
     "./stories/**/*.{js,ts,jsx,tsx}",
+    "./app/**/*.{js,ts,jsx,tsx}",
   ],
   darkMode: "class",
   theme: {